requirements.txt
django
django-registration
django-json-rpc